Moved source_target_file_list to utils/__init__.py (overwrote the deprecated version there).
